+++ jsonpath_rfc9535/node.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 from typing import List
+from typing import Optional
 from typing import Tuple
 from typing import Union
 
@@ -16,14 +17,20 @@ if TYPE_CHECKING:
 class JSONPathNode:
     """A JSON-like value and its location in a JSON document.
 
+    Assigning to `JSONPathNode.value` will update and mutate source data too.
+    Updating data after evaluating a query can invalidate existing child
+    nodes. Use at your own risk.
+
     Attributes:
         value: The JSON-like value at this node.
         location: The names indices that make up the normalized path to _value_.
+        parent: The parent node, or None if this is the root node.
     """
 
     __slots__ = (
-        "value",
+        "_value",
         "location",
+        "parent",
         "root",
     )
 
@@ -32,11 +39,27 @@ class JSONPathNode:
         *,
         value: object,
         location: Tuple[Union[int, str], ...],
+        parent: Optional[JSONPathNode],
         root: JSONValue,
     ) -> None:
-        self.value: object = value
+        self._value: object = value
         self.location: Tuple[Union[int, str], ...] = location
+        self.parent = parent
         self.root = root
+
+    @property
+    def value(self) -> object:
+        """The JSON-like value at this node."""
+        return self._value
+
+    @value.setter
+    def value(self, val: object) -> None:
+        parent = self.parent
+        if parent is not None and self.location:
+            # If data has changed since this node was created, this could fail.
+            # Letting the exception raise is probably the most useful thing we can do.
+            parent._value[self.location[-1]] = val  # type: ignore  # noqa: SLF001
+        self._value = val
 
     def path(self) -> str:
         """Return the normalized path to this node."""
@@ -45,11 +68,17 @@ class JSONPathNode:
             for p in self.location
         )
 
-    def new_child(self, value: object, key: Union[int, str]) -> JSONPathNode:
+    def new_child(
+        self,
+        value: object,
+        key: Union[int, str],
+        parent: Optional[JSONPathNode],
+    ) -> JSONPathNode:
         """Return a new node using this node's location."""
         return JSONPathNode(
             value=value,
             location=self.location + (key,),
+            parent=parent,
             root=self.root,
         )